IN-111 refactor: move logic from GeneralInfoForm into useFillGeneralForm custom hook
