+++ src/features/profile/ui/general-info/GeneralInfoForm.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react'
-import { useForm } from 'react-hook-form'
+import React from 'react'
 
 import CalendarOutline from '@/assets/icons/outlineIcons/CalendarOutline'
 import {
@@ -11,82 +10,22 @@ import {
   TextArea,
   TextField,
 } from '@/common'
-import AuthStore from '@/features/auth/model/authStore'
-import { generalInfoSchema } from '@/features/profile/model/generalInfoSchema'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { format } from 'date-fns'
-import { ru } from 'date-fns/locale'
+import { useFillGeneralForm } from '@/features/profile/model/useFillGeneralForm'
 
-type FormData = {
-  aboutMe?: string
-  city?: string
-  country?: string
-  date?: string
-  firstName: string
-  lastName: string
-  username: string
-}
 export const GeneralInfoForm = React.forwardRef<HTMLFormElement>((_, ref) => {
-  const [toggleOpen, setToggleOpen] = useState<boolean>(false)
-  const calendarRef = useRef<HTMLDivElement>(null)
-  const profile = AuthStore.profile
   const {
+    calendarOpenHandler,
+    calendarRef,
+    city,
     control,
-    formState: {},
+    country,
     handleSubmit,
+    onSubmit,
     register,
+    selectDateHandler,
     setValue,
-    watch,
-  } = useForm<FormData>({
-    defaultValues: {
-      city: '',
-      country: '',
-      date: '',
-      firstName: '',
-      lastName: '',
-      username: profile?.userName,
-    },
-    resolver: zodResolver(generalInfoSchema),
-  })
-
-  const country = watch('country')
-  const city = watch('city')
-
-  const onSubmit = (data: FormData) => {
-    alert(data)
-  }
-
-  const calendarOpenHandler = () => {
-    setToggleOpen(true)
-  }
-
-  const calendarOutsideClickHandler = (e: MouseEvent) => {
-    if (
-      calendarRef.current &&
-      !calendarRef.current.contains(e.target as Node)
-    ) {
-      setToggleOpen(false)
-    }
-  }
-
-  useEffect(() => {
-    if (toggleOpen) {
-      document.addEventListener('mousedown', calendarOutsideClickHandler)
-    } else {
-      document.removeEventListener('mousedown', calendarOutsideClickHandler)
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', calendarOutsideClickHandler)
-    }
-  }, [toggleOpen])
-
-  const selectDateHandler = (date: Date) => {
-    const formattedDate = format(date, 'dd.MM.yyyy', { locale: ru })
-
-    setValue('date', formattedDate)
-    setToggleOpen(false)
-  }
+    toggleOpen,
+  } = useFillGeneralForm()
 
   return (
     <form
@@ -126,7 +65,7 @@ export const GeneralInfoForm = React.forwardRef<HTMLFormElement>((_, ref) => {
           {toggleOpen && (
             <div ref={calendarRef}>
               <Calendar
-                className={'absolute right-0'}
+                className={'absolute'}
                 onDayClick={selectDateHandler}
               />
             </div>
